Move IOC to GPIO module
